requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
python-dotenv
pypdf
unstructured[all-docs]